bookshop/requirements.txt
Django
psycopg2-2.7.6.1
Faker
factory_boy

ipython
